Pack the download request and notification in a structure
